+++ src/plan/store.rs
@@ -1,3 +1,5 @@
+use std::collections::HashMap;
+
 pub struct DataStore {
     store: HashMap<u64, serde_json::Value>,
 }
